WIP: Issue 46
Summary:
- FIXME: commented out all the tests but in the coordinator for ease of
testing
- New tests for coordinator, for location movement + push notification
