+++ imports/api/OCEManager/OCEs/experiences.tests.js
@@ -1,16 +1,16 @@
-import { Experiences, SituationDescription } from './experiences';
-import { CONSTANTS } from "../../Testing/testingconstants";
-
-describe('Experience Tests', () => {
-  it('insert OCEs into database', () => {
-
-    _.forEach(Object.keys(CONSTANTS.EXPERIENCES), (OCE_NAME) => {
-      Experiences.insert(CONSTANTS.EXPERIENCES[OCE_NAME], (err) => {
-        if (err) {
-          chai.assert(false, `Failed to insert CONSTANTS.EXPERIENCED.${OCE_NAME}`);
-        }
-      });
-    });
-
-  })
-});
+// import { Experiences, SituationDescription } from './experiences';
+// import { CONSTANTS } from "../../Testing/testingconstants";
+//
+// describe('Experience Tests', () => {
+//   it('insert OCEs into database', () => {
+//
+//     _.forEach(Object.keys(CONSTANTS.EXPERIENCES), (OCE_NAME) => {
+//       Experiences.insert(CONSTANTS.EXPERIENCES[OCE_NAME], (err) => {
+//         if (err) {
+//           chai.assert(false, `Failed to insert CONSTANTS.EXPERIENCED.${OCE_NAME}`);
+//         }
+//       });
+//     });
+//
+//   })
+// });
