+++ imports/api/UserMonitor/detectors/detectors.tests.js
@@ -1,3 +1,4 @@
+/*
 import { resetDatabase } from 'meteor/xolvio:cleaner';
 import {
   matchAffordancesWithDetector,
